refactor: extract RocketChatGroupClient from RocketChatService
Move group CRUD, membership, and paginated listing operations
(25 methods: createPrivateGroup, deleteGroup*, rollbackGroup,
addUserToGroup, removeUserFromGroup*, getStandardMembersOfGroup,
findMembers, getMembersOfGroup, fetchAllInactivePrivateGroups*,
and private helpers) to new RocketChatGroupClient.

RocketChatGroupClient depends on RocketChatRoomClient for
getChatUsers (MongoDB access). Delegate from RocketChatService
with 1-line calls. Remove all group-related private helpers
and constants from RocketChatService.
